reorg repos collection methods
allows helpers package to use the repos collection
